# swagger/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "REST API Project courses",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://localhost:3000/api"
    }
  ],
...
  "paths": {
    "/roles": {
      "post": {
        "tags": ["roles"],
        "summary": "Add role",
        "description": "Add role",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Role"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          },
          "400": {
            "description": "Role not created"
          },
          "500": {
            "description": "Request failed"
          }
        }
      }
    },
    "/roles/all": {
      "get": {
        "tags": ["roles"],
        "summary": "Get list of roles",
        "description": "Get list of roles",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Roles"
                }
              }
            }
          },
          "404": {
            "description": "Roles not found"
          },
          "500": {
            "description": "Request failed"
          }
        }
      }
    },
    "/roles/{id}": {
      "get": {
        "tags": ["roles"],
        "summary": "Get role by id",
        "description": "Get role by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Role id",
            "schema": {
              "type": "string"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          },
          "404": {
            "description": "Role not found"
          },
          "500": {
            "description": "Request failed"
          }
        }
      },
      "patch": {
        "tags": ["roles"],
        "summary": "Update role by id",
        "description": "Update role by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Role id",
            "schema": {
              "type": "string"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          },
          "404": {
            "description": "Role not found"
          },
          "500": {
            "description": "Request failed"
          }
        }
      },
      "delete": {
        "tags": ["roles"],
        "summary": "Delete role by id",
        "description": "Delete role by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Role id",
            "schema": {
              "type": "string"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {}
            }
          },
          "404": {
            "description": "Role not found"
          },
          "500": {
            "description": "Request failed"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Role": {
        "type": "object",
        "required": [
          "id",
          "name",
          "description"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "enum": [
              "student",
              "teacher",
              "admin"
            ],
            "default": "student"
          },
          "description": {
            "type": "string",
            "example": "student account with rights to view courses"
          }
        }
      },
      "Roles": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Role"
        }
      },
...
    }
  }
}